# use Infinity to improve semantics and performance