--- a/lib/level.js
+++ b/lib/level.js
@@ -1,6 +1,6 @@
 'use strict';
 
-exports.ALL = -100000;
+exports.ALL = -Infinity;
 
 /**
  * Debug log for execute tracing
@@ -31,4 +31,4 @@ exports.WARN = 2;
  */
 exports.ERROR = 3;
 
-exports.NONE = 100000;
+exports.NONE = Infinity;